refactor: 파일 도메인에 따른 패키징
