+++ src/main/java/com/windry/chordplayer/dto/song/CreateSongDto.java
@@ -1,5 +1,6 @@
-package com.windry.chordplayer.dto;
+package com.windry.chordplayer.dto.song;
 
+import com.windry.chordplayer.dto.lyrics.LyricsDto;
 import com.windry.chordplayer.spec.Gender;
 import lombok.Builder;
 import lombok.Data;
