+++ src/main/java/com/windry/chordplayer/dto/song/FiltersOfDetailSong.java
@@ -1,4 +1,4 @@
-package com.windry.chordplayer.dto;
+package com.windry.chordplayer.dto.song;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.windry.chordplayer.spec.Tuning;
